cli/pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "labwatch"
version = "0.1.0"
description = "General-purpose homelab monitoring CLI"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=12.0",
    "docker>=6.0",
    "requests>=2.28",
    "psutil>=5.9",
]

[project.optional-dependencies]
test = ["pytest>=7.0"]

[project.scripts]
labwatch = "labwatch.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
